youusef

+++ src/metier/dao/Implementations/RoleDao.java
@@ -1,12 +1,13 @@
 package metier.dao.Implementations;
 
 import java.util.List;
-import javax.management.relation.Role;
 import org.hibernate.Session;
+
+import metier.dao.beans.Role;
 import metier.dao.util.HibernateUtil;
 import metier.dao.util.functionStd;
 
-public class RoleDao implements functionStd<Role, Long> {
+public class RoleDao implements functionStd<Role, Integer> {
 
 	Session session = HibernateUtil.openSession();
 	private static RoleDao instance = null;
@@ -32,7 +33,7 @@ public class RoleDao implements functionStd<Role, Long> {
 	}
 
 	@Override
-	public void delete(Long b) {
+	public void delete(Integer b) {
 		session.beginTransaction();
 		Role role = getById(b);
 		session.delete(role);
@@ -41,7 +42,7 @@ public class RoleDao implements functionStd<Role, Long> {
 	}
 
 	@Override
-	public Role getById(Long id) {
+	public Role getById(Integer id) {
 		return session.get(Role.class, id);
 	}
